backend-microservices/services/auth-service/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
pymongo==3.12.0
PyJWT==2.8.0
python-dotenv==0.19.0 